archive/requirements.txt
viam-sdk==0.44.0
pygame>=2.5.2,<3.0.0